+++ service.main.js
@@ -1,3 +1,7 @@
+const fs = require("fs-extra");
+const series = require("run-series");
+const forEachItem = require("async-each");
+const contFolders = require("./settings");
 const rioIndex = require("./fox-devices/remote_io/remote-io.index");
 const restartTime = 1000;
 const maxSize = 64000000;
@@ -73,16 +77,79 @@ class Controller
 	
 	factoryReset(factResetCallback)
 	{
-		// Todo
+		var callControllerRestart = this.restartProcess;
+		
 		rioIndex.listRemoteIoDevices(function(deviceQueryErr, deviceQueryRes)
 		{
-			return factResetCallback(deviceQueryErr, deviceQueryRes);
+			if (deviceQueryErr !== null)
+			{
+				return factResetCallback(deviceQueryErr, null);
+			}
+			else
+			{
+				clearDatabase(deviceQueryRes, factResetCallback);
+			}
 		});
+		
+		
+		
+		function clearDatabase(deviceListArray, clearCallback)
+		{
+			forEachItem(deviceListArray,
+			function (currentDevice, currentCallback)
+			{
+				rioIndex.delRemoteIoDevice(currentDevice.id, true, currentCallback);
+			},
+			function (clearErr, clearRes)
+			{
+				if (clearErr !== undefined && clearErr !== null)
+				{
+					return clearCallback(clearErr, null);
+				}
+				else
+				{
+					deleteSaveFiles(clearCallback);
+				}
+			});
+		}
+		
+		
+		
+		function deleteSaveFiles(deleteCallback)
+		{
+			series(
+			[
+				rioIndex.closeRioDatabase.bind(null),
+				fs.remove.bind(null, contFolders.flowsPath),
+				fs.remove.bind(null, contFolders.dbsPath),
+				fs.remove.bind(null, contFolders.userStoragePath),
+				fs.remove.bind(null, contFolders.logsPath)
+			],
+			function (deleteErr, deleteRes)
+			{
+				if (deleteErr !== null)
+				{
+					return deleteCallback(deleteErr, null);
+				}
+				else
+				{
+					callControllerRestart();
+					return deleteCallback(null, true);
+				}
+			});
+		}
 	}
 }
 
 
 const controllerObject = new Controller();
+controllerObject.factoryReset(function(x, y)
+{
+	console.log(x);
+	console.log(y);
+});
+
+
 
 module.exports =
 {